Fixing compilation problem with reloaded pattern.
The system depended upon code which depended upon a custom exception.
Custom exceptions must be compiled. However user.clj was being loaded by
Clojure before Leiningen had a chance to compile the custom exception.
This was a chicken-and-egg problem which was solved by moving the
reloaded pattern functions out of user.clj and into a separate file
which is loaded into the REPL when the REPL starts.